+++ libdrgn/dwarf_info.h
@@ -173,7 +173,7 @@ struct drgn_dwarf_index_state {
 	struct drgn_debug_info *dbinfo;
 	/** Per-thread arrays of CUs to be indexed. */
 	struct drgn_dwarf_index_cu_vector *cus;
-	size_t max_threads;
+	int max_threads;
 };
 
 /**
